app/requirements.txt
Django==4.1.6
django-allauth==0.52.0
django-crispy-forms==1.14.0
gunicorn==20.1.0
django-request==1.6.2
Pillow==9.4.0